Split api.py into two seperate modules for users and parcels
