+++ app/__init__.py
@@ -3,5 +3,5 @@ from config import Config
 
 app = Flask(__name__)
 app.config.from_object(Config)
-from app import api
+from app import api_users, api_parcels
