+++ sendit.py
@@ -1,2 +1,4 @@
 from app import app
 
+if __name__ == '__main__':
+	app.run(debug=True)
